float fix

+++ module_09/ex00/BitcoinExchange.hpp
@@ -6,7 +6,7 @@
 /*   By: cter-maa <cter-maa@student.42.fr>            +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2024/01/15 12:26:16 by chavertterm   #+#    #+#                 */
-/*   Updated: 2024/05/14 15:06:38 by cter-maa      ########   odam.nl         */
+/*   Updated: 2024/05/17 13:22:03 by cter-maa      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -40,7 +40,7 @@ class BitcoinExchange{
 	
 	private:
 		std::map<long, double> _map;
-		int	_value;
+		float	_value;
 		int	_date;
 };
 #endif
